+++ lib/src/elements/filter/mod.rs
@@ -42,9 +42,9 @@ mod video_synchronizer;
 pub use audio::{
     AudioCompressor, AudioCompressorError, AudioCompressorHandle, AudioCompressorOptions,
     AudioGate, AudioGateError, AudioGateHandle, AudioGateOptions, AudioLimiter, AudioLimiterError,
-    AudioLimiterHandle, AudioLimiterOptions, AudioResampler, AudioResamplerError, AudioVolume,
-    AudioVolumeError, AudioVolumeHandle, AudioVolumeOptions, AudioWaveform, AudioWaveformError,
-    AudioWaveformOptions,
+    AudioLimiterHandle, AudioLimiterOptions, AudioResampler, AudioResamplerError, AudioTempo,
+    AudioTempoError, AudioVolume, AudioVolumeError, AudioVolumeHandle, AudioVolumeOptions,
+    AudioWaveform, AudioWaveformError, AudioWaveformOptions,
 };
 #[cfg(feature = "rnnoise")]
 pub use audio::{NOISE_SUPPRESSOR_SAMPLE_RATE, NoiseSuppressor, NoiseSuppressorError};